Remove proposed MenuHelper
Changed:
- Moved methods from `MenuHelper` to `Timber` and replaced occurrences of predecessor.
- Removed filter "timber/menu_helper/menu_locations".
